reset again

+++ kernels/functions.cu
@@ -38,7 +38,7 @@ __host__ __device__ float ActFunc(float x){
     return tanh(x);
 }
 
-__host__ __device__ float scoreFunc(float whenGuess, float whenAns, int hour, float latGuess, float lonGuess, float latAns, float lonAns){
+__host__ __device__ double scoreFunc(double whenGuess, double whenAns, int hour, float latGuess, float lonGuess, float latAns, float lonAns){
     return 1/(fabs(whenGuess-whenAns-hour)*distCalc(latGuess, lonGuess, latAns, lonAns));
 }
